# docs/swagger.yaml
basePath: /
definitions:
...
  ent.Role:
    properties:
      code:
        description: 角色标识
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: 角色描述
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.RoleEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the RoleQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      is_default:
        description: 是否默认角色,默认角色不能删除
        type: boolean
      name:
        description: 角色名称
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  ent.RoleEdges:
    properties:
      users:
        description: Users holds the value of the users edge.
        items:
          $ref: '#/definitions/ent.User'
        type: array
    type: object
...
  ent.User:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      email:
        description: Email holds the value of the "email" field.
        type: string
      email_verified:
        description: EmailVerified holds the value of the "email_verified" field.
        type: boolean
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      phone_number:
        description: PhoneNumber holds the value of the "phone_number" field.
        type: string
      phone_number_verified:
        description: PhoneNumberVerified holds the value of the "phone_number_verified"
          field.
        type: boolean
      role_id:
        description: 角色ID
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  ent.UserEdges:
    properties:
      role:
        allOf:
        - $ref: '#/definitions/ent.Role'
        description: Role holds the value of the role edge.
    type: object
...
  model.HttpError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  model.HttpSuccess:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.PageResult-model_UserResp:
    properties:
      records:
        items:
          $ref: '#/definitions/model.UserResp'
        type: array
      total:
        type: integer
    type: object
  model.PersonalAccessTokenCreateReq:
    properties:
      description:
        description: 描述
        type: string
      expires:
        description: 过期时间
        type: string
      name:
        description: 令牌名称
        type: string
    type: object
  model.PersonalAccessTokenListResp:
    properties:
      description:
        description: 描述
        type: string
      expires:
        description: 过期时间
        type: string
      id:
        type: integer
      name:
        description: 令牌名称
        type: string
    type: object
  model.PersonalAccessTokenResp:
    properties:
      description:
        description: 描述
        type: string
      expires:
        description: 过期时间
        type: string
      id:
        type: integer
      name:
        description: 令牌名称
        type: string
      token:
        description: 令牌
        type: string
    type: object
...
  model.UserProfileResp:
    properties:
      email:
        type: string
      email_verified:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
      phone_number_verified:
        type: boolean
      role:
        $ref: '#/definitions/ent.Role'
      role_id:
        type: integer
      user_id:
        type: integer
    type: object
  model.UserResp:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
      phone_number_verified:
        type: boolean
      role:
        $ref: '#/definitions/ent.Role'
      role_id:
        type: integer
    type: object
...
host: localhost:13000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /api/v1/user/create:
    post:
      consumes:
      - application/json
      description: 创建一个新用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/ent.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建用户
      tags:
      - user
  /api/v1/user/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HttpSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除用户
      tags:
      - user
  /api/v1/user/list:
    get:
      consumes:
      - application/json
      description: 获取所有用户的列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取用户列表
      tags:
      - user
  /api/v1/user/page:
    get:
      consumes:
      - application/json
      description: 获取所有用户的分页列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-model_UserResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取用户分页列表
      tags:
      - user
  /api/v1/user/personal-access-token/create:
    post:
      consumes:
      - application/json
      description: 创建一个新的个人令牌
      parameters:
      - description: 个人令牌创建请求
        in: body
        name: createReq
        required: true
        schema:
          $ref: '#/definitions/model.PersonalAccessTokenCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HttpSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建 personalAccessToken 个人令牌
      tags:
      - user
  /api/v1/user/personal-access-token/list:
    get:
      consumes:
      - application/json
      description: 查询当前用户的所有个人令牌
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.PersonalAccessTokenListResp'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询个人令牌列表
      tags:
      - user
  /api/v1/user/personal-access-token/query/{id}:
    get:
      consumes:
      - application/json
      description: 查询指定个人令牌的详细信息
      parameters:
      - description: 个人令牌ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PersonalAccessTokenResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询个人令牌
      tags:
      - user
  /api/v1/user/profile:
    get:
      consumes:
      - application/json
      description: 查询指定用户的个人信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserProfileResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询用户个人信息
      tags:
      - user
  /api/v1/user/query/{id}:
    get:
      consumes:
      - application/json
      description: 查询指定用户的详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询用户
      tags:
      - user
  /api/v1/user/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定用户的信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/ent.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新用户
      tags:
      - user
swagger: "2.0"